Created FollowingEnemy.java (if santa is confirmed, need to program rotate). Reverse system to float(it is obvious for FollowingEnemy.java)

+++ SquiGoWithSwing/src/Game/Model/BasicTrail.java
@@ -11,7 +11,7 @@ public class BasicTrail extends GameObject {
     float life; //life = 0.001-0.1
 
 
-    public BasicTrail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
+    public BasicTrail(float x, float y, ID id, Color color, int width, int height, float life, Handler handler) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -34,7 +34,7 @@ public class BasicTrail extends GameObject {
         Graphics2D g2d = (Graphics2D) g;
         g2d.setComposite(makeTransparent(alpha));
         g.setColor(color);
-        g.fillRect(x, y, width, height);
+        g.fillRect((int)x, (int)y, width, height);
         //  g2d.setComposite(makeTransparent(1));
 
     }
